% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile.R
\name{avar}
\alias{avar}
\title{Asymptotic variance}
\usage{
avar(w.1, w.0, omega.1, omega.0, T.grad)
}
\arguments{
\item{w.1}{A \code{n_1} by \code{k} dimensional matrix of weight values corresponding to treated observations}

\item{w.0}{A \code{n_0} by \code{k} dimensional matrix of weight values corresponding to control observations}

\item{omega.1}{A \code{n_1} by \code{k} by \code{k} array of conditional variance corresponding to treated observations}

\item{omega.0}{A \code{n_0} by \code{k} by \code{k} array of conditional variance corresponding to control observations}

\item{T.grad}{A \code{k} dimensional gradient vector of \code{T_t f}}
}
\value{
A scalar asymptotic variance value
}
\description{
Calculates the asymptotic variance expression given weights and conditional variance matrix.
}
\examples{
avar(rep(1/50, 50), rep(1/50, 50), rep(1, 50), rep(1, 50), 1)
}
