% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_bw.R
\name{bias_Hol}
\alias{bias_Hol}
\title{Hölder class worst-case bias}
\usage{
bias_Hol(x, t, M, kern, h)
}
\arguments{
\item{x}{a vector of regressors.}

\item{t}{a scalar evaluation point.}

\item{M}{bound on the second derivative}

\item{kern}{a string for kernel name; currently \code{"tri"} is supported.}

\item{h}{a positive bandwidth.}
}
\value{
a scalar worst-case bias value
}
\description{
Calculates the worst-case supremum bias for regression function value estimator
under Hölder class.
}
