% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_bw.R
\name{bias_Lip}
\alias{bias_Lip}
\title{Lipschitz class worst-case bias}
\usage{
bias_Lip(x, t, M, kern, h)
}
\arguments{
\item{x}{a vector of regressors.}

\item{t}{a scalar evaluation point.}

\item{M}{bound on the first derivative}

\item{kern}{a string for kernel name; currently \code{"triangle"} is supported.}

\item{h}{a positive bandwidth.}
}
\value{
a scalar worst-case bias value
}
\description{
Calculates the worst-case bias for regression function values under Lipschitz class
}
