% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_bw.R
\name{bias_Lip_gr}
\alias{bias_Lip_gr}
\title{Lipschitz class worst-case bias: gradient}
\usage{
bias_Lip_gr(x, t, M, kern, h)
}
\arguments{
\item{x}{a vector of regressors.}

\item{t}{a scalar evaluation point.}

\item{M}{bound on the first derivative}

\item{kern}{a string for kernel name; currently \code{"triangle"} is supported.}

\item{h}{a positive bandwidth.}
}
\value{
a scalar worst-case bias gradient value
}
\description{
Calculates the graident for worst-case supremum bias with respect to the bandwidth,
 for regression function value estimator under Lipschitz class
}
