% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_bw.R
\name{bw_Hol_reg}
\alias{bw_Hol_reg}
\title{Optimal bandwidth calculation for regression function estimation under Hölder class}
\usage{
bw_Hol_reg(y, x, eval, C, kern = "triangular", se.initial = "EHW")
}
\arguments{
\item{y}{a vector of dependent variable}

\item{x}{a vector of regressor}

\item{eval}{a vector of evaluation points}

\item{C}{bound on the second derivative}

\item{kern}{kernel function}

\item{se.initial}{initial variance estimation method}
}
\value{
a vector of optimal bandwidths
}
\description{
Optimal bandwidth calculation for regression function estimation under Hölder class
}
\examples{
x <- stats::runif(500, min = -1, max = 1)
y <- x + rnorm(500, 0, 1/4)
eval <- seq(-1, 1, length.out = 10)
bw_Hol_reg(y, x, eval, 1)
}
