% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_bw.R
\name{bw_Lip_supp}
\alias{bw_Lip_supp}
\title{Optimal bandwidths under Lipschitz class with supplied conditional variance}
\usage{
bw_Lip_supp(c.sd, x, t, TE = FALSE, d = NULL, M, kern, alpha, bw.eq = TRUE)
}
\arguments{
\item{c.sd}{a vector of supplied conditional standard deviations}

\item{x}{a vector of regressors.}

\item{t}{a scalar evaluation point.}

\item{TE}{logical specifying whether there are treatment and control groups.}

\item{d}{a vector of indicator variables specifying treatment and control group status;
relevant only when \code{TE = TRUE}.}

\item{M}{bound on the first derivative}

\item{kern}{a string for kernel name; currently \code{"triangle"} is supported.}

\item{alpha}{determines confidence level \code{1 - alpha}.}

\item{bw.eq}{if \code{TRUE}, the same bandwidths are used for estimators for treatment and control groups;
relevant only when \code{TE = TRUE}.}
}
\value{
a list with the following components
\describe{

\item{h.opt}{the optimal bandwidth; when \code{TE = TRUE}, return two bandwidths
for estimators for treatment and control groups.}

\item{hl.opt}{the optimal half-length when the optimal bandwidth(s) is used.}
}
}
\description{
Calculates the optimal bandwidths for regression function value estimator under Lipschitz class
when the conditional variances are known.
}
