% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cb_const.R
\name{cb_const}
\alias{cb_const}
\title{Optimal confidence band}
\usage{
cb_const(
  method,
  C.vec,
  y,
  x,
  d,
  eval,
  T.grad.mat,
  level,
  deg,
  kern,
  M,
  var.reg = "npr",
  seed = NULL,
  useloop = TRUE,
  root.robust = FALSE,
  ng = 10,
  x.out = NULL
)
}
\arguments{
\item{method}{character string used to define the parameter of interest and the function space.
Possible choices are
\describe{
\item{"reg.Hol"}{Regression function values under Hölder space}
\item{"reg.Lip"}{Regression function values under Lipschitz space}
\item{"TE.Lip"}{CATE under Lipschitz space}
\item{"TE.Lip.eqbw"}{CATE under Lipschitz space using the same bandwidths for the treatment
and the control groups}
}}

\item{C.vec}{smoothness parameter for the function space; possibly a vector.}

\item{y}{dependent variable; possibly a matrix with \code{nrow(y)} equals the number of observations.}

\item{x}{a vector of independent variable}

\item{d}{a vector of treatment indicator; it can be arbitrarily specified when not used.}

\item{eval}{a vector of indices}

\item{T.grad.mat}{A \code{n.T} by \code{k} dimensional gradient matrix of \code{T_t f} for \code{t} = 1,..., \code{n.T}.}

\item{level}{level of quantile}

\item{deg}{degree of local polynomial regression to be used}

\item{kern}{kernel used to calculate conditional variance function;
supports \code{"tri"}, \code{"epa"},
\code{"uni"}, and \code{"gau"}. Default is \code{kern = "epa"}.}

\item{M}{number of bootstrap simulations}

\item{var.reg}{nonparametric regression method used to calculate residuals;
either \code{"npr"} (standing for \code{nprobust} package) or \code{"locpol"}
(standing for \code{locpol} package). Default is \code{var.reg = "npr"}.}

\item{seed}{seed for the random number generation; default is \code{seed = NULL}.}

\item{useloop}{If \code{TRUE}, the function is implemented by \code{for} loop over \code{t} = 1,..., \code{n.T};
currently only \code{useloop = TRUE} is implemented.}

\item{root.robust}{if \code{TRUE}, the fuction conducts diagnostic test whether optimizaiton worked well;
default is \code{root.robust = FALSE}.}

\item{ng}{the number of grids of quantile values over which the diagnostic test would be peformed
if \code{root.robust = TRUE}; default is \code{ng = 10}.}

\item{x.out}{the grid of points the confidence band is evaluated; see details.}
}
\value{
a data frame containing index set, corresponding confidence band values, and
bandwidths used for the treatment and the control groups (for models without treatment and control groups,
the same value of bandwidths will be returned). If \code{root.robust = TRUE},
the function returns a list containing the data frame without bandwidths data
 as well as \code{increasing} and \code{opt.grid}; see \code{\link{opt_w}}.
}
\description{
Calculate confidence band by searching for the optimal coverage probability
for individual confidence intervals.
}
\details{
The confidence band is calculated over the points in \code{eval} and interpolated over
\code{x.out} using linear approximation.
}
