% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{K_fun}
\alias{K_fun}
\title{Kernel function}
\usage{
K_fun(x, t, h, kern = c("tri"))
}
\arguments{
\item{x}{a vector of regressors.}

\item{t}{a scalar evaluation point.}

\item{h}{a positive bandwidth.}

\item{kern}{a string for kernel name; currently \code{"tri"} is supported.}
}
\value{
a vector of kernel values with the same dimension as \code{x}.
}
\description{
Calculates the values of kernel function given a vector of regressors.
}
\examples{
x <- seq(-1, 1, length.out = 10)
K_fun(x, 0, 0.5, "tri")
}
