% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{K_gr}
\alias{K_gr}
\title{Kernel gradient function}
\usage{
K_gr(x, t, h, kern = c("triangle"))
}
\arguments{
\item{x}{a vector of regressors.}

\item{t}{a scalar evaluation point.}

\item{h}{a positive bandwidth.}

\item{kern}{a string for kernel name; currently \code{"triangle"} is supported.}
}
\value{
a vector of kernel values with the same dimension as \code{x}.
}
\description{
Calculates the values of the gradient of a given kernel function given a vector of regressors.
Specifically, this function calculates the value of \eqn{\partial K((x - t) / h) / \partial h}.
}
\examples{
x <- seq(-1, 1, length.out = 10)
K_gr(x, 0, 0.5, "triangle")
}
