% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{lp_w}
\alias{lp_w}
\title{Local polynomial regression weigths}
\usage{
lp_w(kern, order, h, x)
}
\arguments{
\item{kern}{name of the kernel (as used in \code{RDHonest} package).}

\item{order}{order of the polynomial regression}

\item{h}{bandwidth}

\item{x}{vector of regressors}
}
\value{
vector of local polynomial regression weights
}
\description{
Local polynomial regression weigths
}
\examples{
lp_w("triangular", 1, 1, seq(-1, 1, length.out = 500))
}
