% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{mat_sq}
\alias{mat_sq}
\title{Columnwise matrix self-multiplication}
\usage{
mat_sq(mat, arr.ret = TRUE)
}
\arguments{
\item{mat}{A matrix or a vector}

\item{arr.ret}{If \code{FALSE}, returns \code{n} by \code{k^2} matrix instead. Default is \code{TRUE}.}
}
\value{
A \code{n} by \code{k} by \code{k} array or a \code{n} by \code{k^2} matrix.
}
\description{
Given a \code{n} by \code{k} matrix \code{mat}, computes a  \code{n} by \code{k} by \code{k} array
whose \code{(i, j1, j2)}th component is given by \code{mat[i, j1] * mat[i, j2]}.
}
\examples{
mat <- matrix(1:12, ncol = 3)
mat_sq(mat, TRUE)
mat_sq(1:10, FALSE)
}
