% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{NpregBand}
\alias{NpregBand}
\title{Confidence band for nonparametric regression}
\usage{
NpregBand(
  y,
  x,
  C,
  level,
  fclass = c("L", "H"),
  n.eval = length(x)/5,
  eval = NULL,
  q.int = 0.025,
  n.sim = 10^3,
  kern = "tri",
  deg = 1,
  var.reg = "npr",
  seed = NULL,
  root.robust = FALSE,
  ng = 10,
  x.out = NULL,
  c.method = "root"
)
}
\arguments{
\item{y}{vector of dependent variables.}

\item{x}{vector of regressors.}

\item{C}{bound on the first or the second derivative, depending on \code{fclass}.}

\item{level}{coverage probability.}

\item{fclass}{function class specification, \code{"L"} for the Lipschitz class, and
\code{"H"} for the Hölder class.}

\item{n.eval}{number of grid points to use when constructing confidence band; default is
\code{n.eval = length(x) / 5}.}

\item{eval}{grid points to use when constructing confidence band; if it is not specified,
\code{eval} is automatically determined by \code{n.eval} and \code{q.int}.}

\item{q.int}{parameter determining the distance from the boundary of the support of \code{x}.
Confidence band is formed for points between the \code{q.int} quantile and the \code{1 - q.int} quantile
of \code{x}. \code{q.int} should be between 0 and 0.5. The default is \code{q.int = 0.025}.}

\item{n.sim}{number of bootstrap samples to use when calculating quantiles of suprema of empirical processes;
the default is \code{n.sim = 10^3}}

\item{kern}{type of kernel used in estimation; currently, only the triangle kernel (\code{kern = "tri"})
is supported.}

\item{deg}{degree of local polynomial estimator used in the first-stage variance estimation;
the default is \code{deg = 1}.}

\item{var.reg}{nonparametric regression method used to calculate residuals;
either \code{"npr"} (standing for \code{nprobust} package) or \code{"locpol"}
(standing for \code{locpol} package). Default is \code{var.reg = "npr"}.}

\item{seed}{seed number for bootstrap random sample generation. The default is \code{seed = NULL}.}

\item{root.robust}{if \code{TRUE}, the fuction conducts diagnostic test whether optimizaiton worked well;
default is \code{root.robust = FALSE}.}

\item{ng}{the number of grids of quantile values over which the diagnostic test would be peformed
if \code{root.robust = TRUE}; default is \code{ng = 10}.}

\item{x.out}{the grid of points the confidence band is evaluated; see details.}

\item{c.method}{method to calculate the optimal value of \eqn{c_n}, an element
in \code{c("root", "supp")}; default is \code{c.method = "root"}.}
}
\value{
a data frame containing index set and corresponding confidence band values,
or if \code{root.robust = TRUE}, a list containing the data frame as well as
\code{increasing} and \code{opt.grid}; see \code{\link{opt_w}}.
}
\description{
Constructs a minimax optimal honest confidence band for nonparametric regression function values
under a specified function class.
}
\details{
Currently, our method is guaranteed to work well for interior points. So setting \code{q.int > 0}
is recommended.
}
\examples{
x <- seq(-1, 1, length.out = 500)
y <- x^2 + rnorm(500, 0, 1/4)
NpregBand(y, x, 2, 0.95, "L", n.eval = 25, c.method = "supp")
NpregBand(y, x, 2, 0.95, "H", n.eval = 25, c.method = "supp")
}
