% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{opt_cn}
\alias{opt_cn}
\title{Optimal value of \eqn{c_n}}
\usage{
opt_cn(n, p)
}
\arguments{
\item{n}{sample size}

\item{p}{Holder space exponent, either 1 or 2.}
}
\value{
the value of the optimal \eqn{c_n}.
}
\description{
Optimal value of \eqn{c_n}
}
\examples{
opt_cn(500, 2)
}
