% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimaztion.R
\name{opt_w}
\alias{opt_w}
\title{Optimal weights}
\usage{
opt_w(
  method,
  C.vec,
  y,
  x,
  d = NULL,
  eval,
  T.grad.mat,
  level,
  deg,
  kern,
  M,
  seed = NULL,
  useloop = TRUE,
  root.robust = FALSE,
  ng = 10,
  resid.1 = NULL,
  resid.0 = NULL,
  var.reg = "npr",
  c.method = "root",
  c.supp = NULL
)
}
\arguments{
\item{method}{character string used to define the parameter of interest and the function space.
Possible choices are
\describe{
\item{"reg.Hol"}{Regression function values under Hölder space}
\item{"reg.Lip"}{Regression function values under Lipschitz space}
\item{"TE.Lip"}{CATE under Lipschitz space}
\item{"TE.Lip.eqbw"}{CATE under Lipschitz space using the same bandwidths for the treatment
and the control groups}
\item{"TE.Lip"}{CATE under Hölder space}
\item{"TE.Lip.eqbw"}{CATE under Hölder space using the same bandwidths for the treatment
and the control groups}
}}

\item{C.vec}{smoothness parameter for the function space; possibly a vector.}

\item{y}{dependent variable; possibly a matrix with \code{nrow(y)} equals the number of observations.}

\item{x}{a vector of independent variable}

\item{d}{a vector of treatment indicator; it can be arbitrarily specified when not used.}

\item{eval}{a vector of indices}

\item{T.grad.mat}{A \code{n.T} by \code{k} dimensional gradient matrix of \code{T_t f} for \code{t} = 1,..., \code{n.T}.}

\item{level}{level of quantile}

\item{deg}{degree of local polynomial regression to be used}

\item{kern}{kernel used to calculate conditional variance function;
supports \code{"tri"}, \code{"epa"},
\code{"uni"}, and \code{"gau"}. Default is \code{kern = "epa"}.}

\item{M}{number of bootstrap simulations}

\item{seed}{seed for the random number generation; default is \code{seed = NULL}.}

\item{useloop}{If \code{TRUE}, the function is implemented by \code{for} loop over \code{t} = 1,..., \code{n.T};
currently only \code{useloop = TRUE} is implemented.}

\item{root.robust}{if \code{TRUE}, the fuction conducts diagnostic test whether optimizaiton worked well;
default is \code{root.robust = FALSE}.}

\item{ng}{the number of grids of quantile values over which the diagnostic test would be peformed
if \code{root.robust = TRUE}; default is \code{ng = 10}.}

\item{resid.1}{residuals corresponding to the treated observations; it can be a vector
or a matrix if \code{ncol(y) > 1}. For treatment effect models, its \eqn{i}th component
corresponds to the \eqn{i}th component of \code{y[d == 1, ]}. If there are no
treatment and control groups, \code{resid.1} corresponds to residuals for the
entire observations. It can be specified to be \code{NULL}.}

\item{resid.0}{residuals corresponding to the control observations; it can be a vector
or a matrix if \code{ncol(y) > 1}. Its \eqn{i}th component
corresponds to the \eqn{i}th component of \code{y[d == 0, ]}. It can be specified to be
\code{NULL}.}

\item{var.reg}{nonparametric regression method used to calculate residuals;
either \code{"npr"} (standing for \code{nprobust} package) or \code{"locpol"}
(standing for \code{locpol} package). Default is \code{var.reg = "npr"}.}

\item{c.method}{method to calculate the optimal value of \eqn{c_n}, an element
in \code{c("root", "supp")}; default is \code{c.method = "root"}.}

\item{c.supp}{value of \eqn{c_n} to be used when \code{c.method = "supp"};
default is \code{c.supp = NULL}.}
}
\value{
a list of the following components
\describe{
\item{w.1}{a 3-dim array of weigths corresponding to the treated observations;
the second and third dimension correponds to \code{ncol(y)} and \code{length(eval)}}
\item{w.0}{a 3-dim array of weigths corresponding to the control observations;
the second and third dimension correponds to \code{ncol(y)} and \code{length(eval)}.
If there is no control observations, returns array of 0.}
\item{c.root}{the quantile value corresponding to the optimal weigths.}
\item{increasing}{boolean value testing whether the objective function is increasing or not;
provided only if \code{root.robust = TRUE}}
\item{opt.grid}{a \code{data.frame} object containing evaluations of the objective function
over grid values of quantiles; provided only if \code{root.robust = TRUE}}
}
}
\description{
Solve for optimal weigths by the quantile matching method
}
