% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimaztion.R
\name{opt_w_2}
\alias{opt_w_2}
\title{Optimal weights (time checking purpose)}
\usage{
opt_w_2(
  c,
  method,
  C.vec,
  y,
  x,
  d,
  eval,
  T.grad.mat,
  level,
  deg,
  kern,
  loo,
  M,
  seed = NULL,
  useloop = TRUE,
  root.robust = FALSE,
  ng = 10
)
}
\arguments{
\item{c}{value of c}

\item{method}{character string used to define the parameter of interest and the function space.
Possible choices are
\describe{
\item{"reg.Hol"}{Regression function values under Hölder space}
\item{"reg.Lip"}{Regression function values under Lipschitz space}
\item{"TE.Lip"}{CATE under Lipschitz space}
\item{"TE.Lip.eqbw"}{CATE under Lipschitz space using the same bandwidths for the treatment
and the control groups}
}}

\item{C.vec}{smoothness parameter for the function space; possibly a vector.}

\item{y}{dependent variable; possibly a matrix with \code{nrow(y)} equals the number of observations.}

\item{x}{a vector of independent variable}

\item{d}{a vector of treatment indicator; it can be arbitrarily specified when not used.}

\item{eval}{a vector of indices}

\item{T.grad.mat}{A \code{n.T} by \code{k} dimensional gradient matrix of \code{T_t f} for \code{t} = 1,..., \code{n.T}.}

\item{level}{level of quantile}

\item{deg}{degree of local polynomial regression to be used}

\item{kern}{name of the kernel. Supports \code{"triangle"} and \code{"epa"}.}

\item{loo}{If \code{TRUE}, the leave-one-out version is used; see \code{\link[locpol]{locPolSmootherC}}.}

\item{M}{number of bootstrap simulations}

\item{seed}{seed for the random number generation; default is \code{seed = NULL}.}

\item{useloop}{If \code{TRUE}, the function is implemented by \code{for} loop over \code{t} = 1,..., \code{n.T};
currently only \code{useloop = TRUE} is implemented.}

\item{root.robust}{if \code{TRUE}, the fuction conducts diagnostic test whether optimizaiton worked well;
default is \code{root.robust = FALSE}.}

\item{ng}{the number of grids of quantile values over which the diagnostic test would be peformed
if \code{root.robust = TRUE}; default is \code{ng = 10}.}
}
\value{
a list of the following components
\describe{
\item{w.1}{a 3-dim array of weigths corresponding to the treated observations;
the second and third dimension correponds to \code{ncol(y)} and \code{length(eval)}}
\item{w.0}{a 3-dim array of weigths corresponding to the control observations;
the second and third dimension correponds to \code{ncol(y)} and \code{length(eval)}.
If there is no control observations, returns array of 0.}
\item{c.root}{the quantile value corresponding to the optimal weigths.}
\item{increasing}{boolean value testing whether the objective function is increasing or not;
provided only if \code{root.robust = TRUE}}
\item{opt.grid}{a \code{data.frame} object containing evaluations of the objective function
over grid values of quantiles; provided only if \code{root.robust = TRUE}}
}
}
\description{
Solve for optimal weigths by the quantile matching method
}
