% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{RDH_lee}
\alias{RDH_lee}
\title{RDHonest result for Lee dataset}
\usage{
RDH_lee()
}
\value{
point estimate for RDHonest procedure on Lee dataset
}
\description{
Return point estimate for RDHonest procedure on Lee dataset.
}
\details{
This function was made for a test purpose for package import.
}
\examples{
RDH_lee()
}
