% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cond_var.R
\name{resid_calc}
\alias{resid_calc}
\title{Residual calculation function}
\usage{
resid_calc(y, x, d = NULL, deg, kern)
}
\arguments{
\item{y}{dependent variable; possibly a matrix with \code{nrow(y)} equals the number of observations.}

\item{x}{a vector of independent variable}

\item{d}{a vector of treatment indicator; it can be arbitrarily specified when not used.}

\item{deg}{degree of local polynomial regression to be used}

\item{kern}{kernel used to calculate conditional variance function;
supports \code{"tri"}, \code{"epa"},
\code{"uni"}, and \code{"gau"}. Default is \code{kern = "epa"}.}
}
\value{
a list of residuals corresponding to treated individuals (\code{resid.1}) and control individuals (\code{resid.0});
if there is no control group, \code{resid.0 = 0} is returned.
}
\description{
Calculates residuals corresponding to each observations.
}
