% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_bw.R
\name{sd_Lip_gr}
\alias{sd_Lip_gr}
\title{Lipschitz class standard deviation: gradient}
\usage{
sd_Lip_gr(y, x, t, kern, h, deg, loo, sd.homo = TRUE)
}
\arguments{
\item{y}{vector of dependent variables}

\item{x}{a vector of regressors.}

\item{t}{a scalar evaluation point.}

\item{kern}{a string for kernel name; currently \code{"triangle"} is supported.}

\item{h}{a positive bandwidth.}

\item{deg}{degree of local polynomial regression to be used}

\item{loo}{If \code{TRUE}, the leave-one-out version is used; see \code{\link[locpol]{locPolSmootherC}}.}

\item{sd.homo}{logical indicating whether the variance would be estimated under locally homoskedastic variance
assumption; the default is \code{sd.homo = TRUE}.}
}
\value{
a scalar standard deviation gradient value
}
\description{
Calculates the gradient of the standard devation with respect to the bandwidth,
 for regression function value estimator under Lipschitz class
}
