% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile.R
\name{sup_quant_orc}
\alias{sup_quant_orc}
\title{True quantile value approximation}
\usage{
sup_quant_orc(
  eps.1.mat,
  eps.0.mat,
  w.1.arr,
  w.0.arr,
  omega.1,
  omega.0,
  T.grad.mat,
  level,
  M,
  useloop = TRUE
)
}
\arguments{
\item{eps.1.mat}{a \code{n.1 * M} by \code{k} matrix of residuals generated from the true distribution,
corresponding to the treated observations}

\item{eps.0.mat}{a \code{n.0 * M} by \code{k} matrix of residuals generated from the true distribution,
corresponding to the treated observations}

\item{w.1.arr}{A \code{n_1} by \code{k} by \code{n.T} dimensional array of weight values corresponding to treated observations}

\item{w.0.arr}{A \code{n_0} by \code{k} by \code{n.T} dimensional array of weight values corresponding to treated observations}

\item{omega.1}{A \code{n_1} by \code{k} by \code{k} array of conditional variance corresponding to treated observations}

\item{omega.0}{A \code{n_0} by \code{k} by \code{k} array of conditional variance corresponding to control observations}

\item{T.grad.mat}{A \code{n.T} by \code{k} dimensional gradient matrix of \code{T_t f} for \code{t} = 1,..., \code{n.T}.}

\item{level}{level of quantile}

\item{M}{number of bootstrap simulations}

\item{useloop}{If \code{TRUE}, the function is implemented by \code{for} loop over \code{t} = 1,..., \code{n.T};
currently only \code{useloop = TRUE} is implemented.}
}
\value{
a scalar quantile value
}
\description{
Calulates the true quantile of the supremum of absolute value of a studentized process
by simulation
}
