% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test_fun}
\alias{test_fun}
\title{Add two numbers}
\usage{
test_fun(a, b)
}
\arguments{
\item{a}{first number}

\item{b}{second number}
}
\value{
the sum of two numbers
}
\description{
Calculate the sum of two numbers
}
\examples{
test_fun(1, 3)
}
