% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{v_to_3d}
\alias{v_to_3d}
\title{Vector to 3-dim array}
\usage{
v_to_3d(v, obj.name)
}
\arguments{
\item{v}{a vector or an array}

\item{obj.name}{a string used to display the error message}
}
\value{
a list containing the following components
\describe{

  \item{\code{res.arr}}{Corresponding 3-dimensional array. If the input is not right, this contains \code{NULL}}

  \item{\code{err.stat}}{Error status variable, taking 1 if the input it not right. Otherwise, this equals 0.}

  \item{\code{err.msg}}{Error message to display if \code{err.stat = 1}. This contains an empty string if \code{err.stat = 0}.}

}
}
\description{
Test whether an input is a vector, and if so, transform it into a array with its second and third dimension lengths equal to 1.
 If an input is neither a vector nor a 3-dimensional array, returns \code{NULL} with an error message.
}
\examples{

v_to_3d(1:10, "omega")
v_to_3d(array(1:10, dim = c(10, 1, 1)), "omega")
v_to_3d(matrix(1:10, ncol = 1), "omega")
}
