% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{v_to_m}
\alias{v_to_m}
\title{Vector to matrix}
\usage{
v_to_m(v)
}
\arguments{
\item{v}{a matrix or a vector}
}
\value{
a corresponding matrix
}
\description{
Test whether an input is a vector, and if so, transform it into a matrix with \code{ncol = 1}.
}
\examples{
v_to_m(1:10)
v_to_m(matrix(1:10, ncol = 1))
}
