% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_bw.R
\name{var_Hol}
\alias{var_Hol}
\title{Hölder class variance}
\usage{
var_Hol(y, x, t, kern, h)
}
\arguments{
\item{y}{vector of dependent variables}

\item{x}{a vector of regressors.}

\item{t}{a scalar evaluation point.}

\item{kern}{a string for kernel name; currently \code{"tri"} is supported.}

\item{h}{a positive bandwidth.}
}
\value{
a scalar variance value
}
\description{
Calculates the variance for regression function value estimator under Hölder class
}
