% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_bw.R
\name{var_Lip}
\alias{var_Lip}
\title{Lipschitz class variance}
\usage{
var_Lip(y, x, t, kern, h, deg, sd.homo = TRUE)
}
\arguments{
\item{y}{vector of dependent variables}

\item{x}{a vector of regressors.}

\item{t}{a scalar evaluation point.}

\item{kern}{a string for kernel name; currently \code{"tri"} is supported.}

\item{h}{a positive bandwidth.}

\item{deg}{degree of local polynomial regression to be used}

\item{sd.homo}{logical indicating whether the variance would be estimated under locally homoskedastic variance
assumption; the default is \code{sd.homo = TRUE}.}
}
\value{
a scalar variance value
}
\description{
Calculates the variance for regression function value estimator under Lipschitz class
}
