% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_bw.R
\name{var_Lip_2}
\alias{var_Lip_2}
\title{Lipschitz class variance (stable version)}
\usage{
var_Lip_2(y, x, t, kern, h, deg, loo)
}
\arguments{
\item{y}{vector of dependent variables}

\item{x}{a vector of regressors.}

\item{t}{a scalar evaluation point.}

\item{kern}{a string for kernel name; currently \code{"triangle"} is supported.}

\item{h}{a positive bandwidth.}

\item{deg}{degree of local polynomial regression to be used}

\item{loo}{If \code{TRUE}, the leave-one-out version is used; see \code{\link[locpol]{locPolSmootherC}}.}
}
\value{
a scalar variance value
}
\description{
Calculates the variance for regression function value estimator under Lipschitz class
}
