% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_bw.R
\name{var_Lip_resid}
\alias{var_Lip_resid}
\title{Lipschitz class variance using residuals (obsolete now)}
\usage{
var_Lip_resid(x, t, kern, h, resid)
}
\arguments{
\item{x}{a vector of regressors.}

\item{t}{a scalar evaluation point.}

\item{kern}{a string for kernel name; currently \code{"tri"} is supported.}

\item{h}{a positive bandwidth.}

\item{resid}{a vector of true conditional standard deviation values}
}
\value{
a scalar variance value
}
\description{
Calculates the variance for regression function value estimator under Lipschitz class using residuals
}
\details{
When \code{resid} corresponds to the true conditional standard deviations, this function calculates
the true variance value.
}
