% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_bw.R
\name{var_Lip_true}
\alias{var_Lip_true}
\title{Lipschitz class true variance}
\usage{
var_Lip_true(x, t, kern, h, sd.true)
}
\arguments{
\item{x}{a vector of regressors.}

\item{t}{a scalar evaluation point.}

\item{kern}{a string for kernel name; currently \code{"triangle"} is supported.}

\item{h}{a positive bandwidth.}

\item{sd.true}{a vector of true conditional standard deviation values}
}
\value{
a scalar variance value
}
\description{
Calculates the true variance for regression function value estimator under Lipschitz class
}
