% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{w_get_Hol}
\alias{w_get_Hol}
\title{Optimal weights for Hölder space}
\usage{
w_get_Hol(
  y,
  x,
  eval,
  M,
  level,
  kern = "triangular",
  se.initial = "EHW",
  se.method = "nn",
  J = 3
)
}
\arguments{
\item{y}{vector of dependent variables}

\item{x}{vector of regressors}

\item{eval}{evaluation points}

\item{M}{bound on the second derivative}

\item{level}{confidence level}

\item{kern}{specifies kernel function used in the local regression; default = \code{"triangular"}.
See \code{\link[RDHonest]{NPROptBW.fit}} in \code{RDHonest} package for a list of kernels available.}

\item{se.initial}{method for estimating initial variance for computing optimal bandwidt; default = \code{"EHW"}.
See \code{\link[RDHonest]{NPROptBW.fit}} in \code{RDHonest} package for a list of method available.}

\item{se.method}{methods for estimating standard error of estimate; default = \code{"nn"}.
See \code{\link[RDHonest]{NPRreg.fit}} in \code{RDHonest} package for a list of method available.}

\item{J}{number of nearest neighbors, if "nn" is specified in se.method.}
}
\value{
list with components

\describe{
    \item{w.mat}{matrix of optimal weigths with a dimension \code{length(y)} by
    \code{length(eval)} corresponding to the evaluation points}
    \item{bw.vec}{vector of optimal bandwidths corresponding to the evaluation points}

}
}
\description{
Computes optimal weights and bandwidths for inference for nonparametric regression function values
under Hölder space, with an aid of \code{RDHonest} package.
}
\examples{
x <- stats::runif(500, min = -1, max = 1)
y <- x + rnorm(500, 0, 1/4)
eval <- seq(from = -0.9, to = 0.9, length.out = 5)
w_get_Hol(y, x, eval, 1, 0.95)
}
