% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{w_get_Lip}
\alias{w_get_Lip}
\title{Optimal weights for Lipschitz space}
\usage{
w_get_Lip(
  y,
  x,
  eval,
  C,
  level,
  TE = FALSE,
  d = NULL,
  kern = "triangle",
  bw.eq = TRUE,
  deg = 0
)
}
\arguments{
\item{y}{vector of dependent variables}

\item{x}{a vector of regressors.}

\item{eval}{evaluation points}

\item{C}{bound on the second derivative}

\item{level}{confidence level of each one-sided confidence intervals}

\item{TE}{logical specifying whether there are treatment and control groups.}

\item{d}{a vector of indicator variables specifying treatment and control group status;
relevant only when \code{TE = TRUE}.}

\item{kern}{a string for kernel name; currently \code{"triangle"} is supported.}

\item{bw.eq}{if \code{TRUE}, the same bandwidths are used for estimators for treatment and control groups;
relevant only when \code{TE = TRUE}.}

\item{deg}{degree of local polynomial regression to be used}
}
\value{
a matrix of weights with each column corresponding to weights for each evaluation points,
or if \code{TE = TRUE}, a list of two such matrices, with \code{w.mat.1} corresponding to that of the
treated group and \code{w.mat.0} corresponding to that of the control group.
}
\description{
Computes optimal weights and bandwidths for inference for nonparametric regression function values
under Lipschitz space.
}
\examples{
x <- stats::runif(500, min = -1, max = 1)
y <- x + rnorm(500, 0, 1/4)
eval <- seq(from = -0.9, to = 0.9, length.out = 5)
w_get_Lip(y, x, eval, 1, 0.95)
}
