% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cond_var.R
\name{eps_hat}
\alias{eps_hat}
\title{Residual calculation}
\usage{
eps_hat(y, x, deg, kern = "epa", var.reg = "npr", n.max = 500)
}
\arguments{
\item{y}{vector of dependent variables}

\item{x}{vector of regressors}

\item{deg}{degree of local polynomial regression to be used}

\item{kern}{kernel used to calculate conditional variance function;
supports \code{"tri"}, \code{"epa"},
\code{"uni"}, and \code{"gau"}. Default is \code{kern = "epa"}.}

\item{var.reg}{nonparametric regression method used to calculate residuals;
either \code{"npr"} (standing for \code{nprobust} package) or \code{"locpol"}
(standing for \code{locpol} package). Default is \code{var.reg = "npr"}.}

\item{n.max}{maximum number of observations allowed where the regression function estimator
is evaluated when \code{var.reg = "npr"}.
If \code{length(x) > n.max}, the regression function estimator
is evaluated over a grid of the length \code{n.max}, and the regression function estimate over
\code{x} is obtained by linear interpolation. The default is \code{n.max = 500}.}
}
\value{
vector of residuals with the same length as \code{y}
}
\description{
Calculates residuals after nonparametric regression using \code{\link[nprobust]{lprobust}}
or \code{\link[locpol]{locPolSmootherC}}.
}
\details{
For now, only \code{kern = "tri"} and \code{kern = "epa"} are supported when
\code{var.reg = "locpol"}.
}
\examples{
x <- seq(from = -1, to = 1, length.out = 500)
y <- x^2 + stats::rnorm(500, 0, 0.1)
eps_hat(y, x, 1)
eps_hat(y, x, 1, var.reg = "locpol")
}
